# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Flappy Bird Backend API
  version: 1.0.0
  summary: Optional accounts, leaderboards, achievements and score sync for the Flappy Bird iOS game.
  description: |
    # Overview

    This service is the **optional** companion backend for the
    [Flappy Bird iOS game](https://github.com/hoangsonww/Flappy-Bird-Game).

    The game is fully playable without it. When a server is reachable, the app
    unlocks accounts, cloud-synced progress, global/daily/weekly leaderboards,
    friend boards, achievements and the deterministic daily challenge.

    ## Discovery handshake

    On launch the game probes a short list of candidate base URLs and calls
    `GET /v1/meta/config`. A response containing `"protocol": "flappy-bird/1"`
    switches the app into online mode; anything else keeps it offline. Nothing
    else about the app's behaviour depends on the server being present.

    ```mermaid
    sequenceDiagram
      participant Game as iOS Game
      participant API as Backend
      Game->>API: GET /v1/meta/config
      API-->>Game: 200 { protocol, capabilities, limits }
      Game->>API: POST /v1/auth/guest  { deviceId }
      API-->>Game: 201 { user, tokens }
      Game->>API: POST /v1/scores      (Bearer access token)
      API-->>Game: 201 { score, rank, unlockedAchievements }
    ```

    ## Authentication

    Two token types:

    | Token         | Lifetime            | Storage on device | Purpose                        |
...

...

...
    history.

...

...

...
    {
...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      description: Creates a password account and immediately returns a token pair.
      security: []
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
            examples:
              minimal:
                summary: Username and password only
                value: { username: skyhopper, password: correct-horse-battery }
              full:
                summary: With profile details
                value:
                  username: skyhopper
                  password: correct-horse-battery
                  email: sky@example.com
                  displayName: Sky Hopper
                  country: US
                  avatarSkin: midnight
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            examples:
              default:
                value: { username: skyhopper, password: correct-horse-battery }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '401':
          $ref: '#/components/responses/InvalidCredentials'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/auth/guest:
    post:
      tags: [Auth]
      summary: Create or resume a guest account
      description: |
        Device-bound account with no password. Calling this twice with the same
        `deviceId` resumes the existing account (`200`) instead of creating a
        duplicate (`201`). Disabled when `ENABLE_GUEST_ACCOUNTS=false`.
      security: []
      operationId: guestLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GuestRequest' }
            examples:
              default:
                value: { deviceId: 6E2A9C71-4F0B-4D2A-9C1E-8F0C2B7A5D31, country: US }
      responses:
        '200':
          description: Existing guest account resumed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '201':
          description: Guest account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate tokens
      description: Consumes a refresh token and returns a brand-new pair. The consumed token is revoked.
      security: []
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string, minLength: 16, maxLength: 512 }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke one refresh token
      description: Idempotent. Omitting the body succeeds and revokes nothing.
      security: []
      operationId: logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '204': { description: Token revoked (or nothing to revoke) }
  /v1/auth/logout-all:
    post:
      tags: [Auth]
      summary: Revoke every session
      description: Signs the account out of all devices.
      operationId: logoutAll
      responses:
        '200':
          description: Number of revoked refresh tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked: { type: integer, example: 3 }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Current account
      operationId: getCurrentAccount
      responses:
        '200':
          description: Account, aggregate stats and active session count
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/PrivateUser' }
                  stats: { oneOf: [{ $ref: '#/components/schemas/UserStats' }, { type: 'null' }] }
                  activeSessions: { type: integer, example: 2 }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/password:
    patch:
      tags: [Auth]
      summary: Change password
      description: Revokes every refresh token on success, forcing a fresh login elsewhere.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '204': { description: Password changed }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/InvalidCredentials'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/auth/upgrade:
    post:
      tags: [Auth]
      summary: Promote a guest to a full account
      description: Keeps every score, stat and achievement already attached to the guest.
      operationId: upgradeGuest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { $ref: '#/components/schemas/Username' }
                password: { type: string, minLength: 8, maxLength: 128 }
                email: { type: [string, 'null'], format: email }
                displayName: { type: [string, 'null'], maxLength: 40 }
      responses:
        '200':
          description: Account upgraded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/auth/account:
    delete:
      tags: [Auth]
      summary: Delete the account
      description: Irreversible. Cascades to runs, stats, achievements, friendships and tokens.
      operationId: deleteAccount
      responses:
        '204': { description: Account deleted }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/challenges/{date}:
    get:
      tags: [Challenges]
      summary: Challenge for a specific date
      security: []
      operationId: getChallengeByDate
      parameters:
        - $ref: '#/components/parameters/DatePath'
      responses:
        '200':
          description: Challenge
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge: { $ref: '#/components/schemas/DailyChallenge' }
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /v1/challenges/{date}/leaderboard:
    get:
      tags: [Challenges]
      summary: Challenge leaderboard
      security: []
      operationId: getChallengeLeaderboard
      parameters:
        - $ref: '#/components/parameters/DatePath'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Ranked entries for that date
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string, format: date }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ChallengeEntry' }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  hasMore: { type: boolean }
  /v1/challenges/today/entries:
    post:
      tags: [Challenges]
      summary: Submit a challenge score
      description: Keeps the best score per player per day.
      operationId: submitTodayChallengeEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [score]
              properties:
                score: { type: integer, minimum: 0 }
      responses:
        '201':
          description: Stored entry (existing higher score is preserved)
          content:
            application/json:
              schema:
                type: object
                properties:
                  entry: { $ref: '#/components/schemas/ChallengeEntry' }
  /v1/challenges/{date}/entries:
    post:
      tags: [Challenges]
      summary: Submit a challenge score for a date
      description: Only today's date is accepted; past dates return `400`.
      operationId: submitChallengeEntry
      parameters:
        - $ref: '#/components/parameters/DatePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [score]
              properties:
                score: { type: integer, minimum: 0 }
      responses:
        '201':
          description: Stored entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  entry: { $ref: '#/components/schemas/ChallengeEntry' }
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token from `/v1/auth/login`, `/register`, `/guest` or `/refresh`.
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Static token matching the server's `ADMIN_TOKEN`. An alternative to an admin account.
  parameters:
...
    DatePath:
      name: date
      in: path
      required: true
      description: UTC date in `YYYY-MM-DD` form.
      schema: { type: string, format: date, example: '2026-03-19' }
...
    LimitQuery:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    OffsetQuery:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  schemas:
    Username:
      type: string
      pattern: '^[a-zA-Z0-9_]{3,20}$'
      example: skyhopper
      description: 3–20 characters; letters, digits and underscore. Compared case-insensitively.
    BirdSkin:
      type: string
      enum: [classic, midnight, ember, mint, royal, glitch, aurora, phoenix]
      description: Cosmetic bird skin. Unlocked in game with coins.
    GameMode:
      type: string
      enum: [classic, endless, timeAttack, hardcore, zen, daily]
      description: |
        * `classic` — the original rules.
        * `endless` — difficulty ramps continuously.
        * `timeAttack` — score as much as possible in 60 seconds.
        * `hardcore` — narrow gaps, one life, no power-ups.
        * `zen` — no death, for practice (excluded from most boards by convention).
        * `daily` — today's deterministic challenge.
...
    PublicUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { $ref: '#/components/schemas/Username' }
        displayName: { type: string, example: Sky Hopper }
        country: { type: [string, 'null'], example: US }
        avatarSkin: { $ref: '#/components/schemas/BirdSkin' }
        isGuest: { type: boolean }
        createdAt: { type: string, format: date-time }
    PrivateUser:
      allOf:
        - $ref: '#/components/schemas/PublicUser'
        - type: object
          properties:
            email: { type: [string, 'null'], format: email }
            role: { type: string, enum: [player, admin] }
            lastSeenAt: { type: string, format: date-time }
    TokenPair:
      type: object
      properties:
        accessToken: { type: string, description: JWT for the `Authorization` header. }
        refreshToken: { type: string, description: Opaque rotating token. Store in the Keychain. }
        tokenType: { type: string, enum: [Bearer] }
        expiresIn: { type: integer, description: Access-token lifetime in seconds., example: 900 }
        refreshExpiresAt: { type: string, format: date-time }
    AuthSession:
      type: object
      properties:
        user: { $ref: '#/components/schemas/PrivateUser' }
        tokens: { $ref: '#/components/schemas/TokenPair' }
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username: { $ref: '#/components/schemas/Username' }
        password: { type: string, minLength: 8, maxLength: 128 }
        email: { type: [string, 'null'], format: email }
        displayName: { type: [string, 'null'], maxLength: 40 }
        country: { type: [string, 'null'], minLength: 2, maxLength: 2 }
        avatarSkin: { $ref: '#/components/schemas/BirdSkin' }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    GuestRequest:
      type: object
      required: [deviceId]
      properties:
        deviceId:
          type: string
          minLength: 8
          maxLength: 128
          description: Stable per-install identifier. The game uses `identifierForVendor`.
        country: { type: [string, 'null'], minLength: 2, maxLength: 2 }
...
    UserStats:
      type: object
      properties:
        userId: { type: string, format: uuid }
        bestScore: { type: integer }
        bestScoreMode: { oneOf: [{ $ref: '#/components/schemas/GameMode' }, { type: 'null' }] }
        totalScore: { type: integer }
        gamesPlayed: { type: integer }
        totalCoins: { type: integer }
        totalPipes: { type: integer }
        totalDurationMs: { type: integer }
        longestRunMs: { type: integer }
        bestCombo: { type: integer }
        updatedAt: { type: string, format: date-time }
...
    DailyChallenge:
      type: object
      properties:
        date: { type: string, format: date }
        seed: { type: string, example: 9f3c0e1187ab2d40 }
        mode: { $ref: '#/components/schemas/GameMode' }
        pipeGap:
          { type: integer, description: Vertical gap between pipes, in points., example: 140 }
        gravityScale: { type: number, example: 1.05 }
        speedScale: { type: number, example: 1.1 }
        modifier: { type: string, enum: [none, windy, foggy, nightfall, narrow, turbo] }
        description: { type: string }
        createdAt: { type: string, format: date-time }
    ChallengeEntry:
      type: object
      properties:
        rank: { type: integer }
        date: { type: string, format: date }
        userId: { type: string, format: uuid }
        username: { $ref: '#/components/schemas/Username' }
        displayName: { type: string }
        avatarSkin: { $ref: '#/components/schemas/BirdSkin' }
        score: { type: integer }
        submittedAt: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - validation_failed
                - unauthorized
                - invalid_credentials
                - token_expired
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - payload_too_large
                - unprocessable
                - service_unavailable
                - internal_error
            message: { type: string }
            details: { description: Structured context; shape depends on `code`. }
            requestId: { type: string, description: Matches the `X-Request-Id` response header. }
  responses:
    BadRequest:
      description: The request was malformed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing, invalid or expired access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    InvalidCredentials:
      description: Username or password is incorrect
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Conflict:
      description: Uniqueness or state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationFailed:
      description: Request failed schema validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    RateLimited:
      description: Too many requests
      headers:
        RateLimit-Limit: { schema: { type: integer } }
        RateLimit-Remaining: { schema: { type: integer } }
        RateLimit-Reset: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }